Move X509Signer from ceremony to pkcs11helpers.
Precursor to #4992.
