+++ pkcs11helpers/helpers_test.go
@@ -1,7 +1,15 @@
 package pkcs11helpers
 
 import (
+	"bytes"
+	"crypto"
+	"crypto/ecdsa"
+	"crypto/elliptic"
+	"crypto/rand"
+	"crypto/sha256"
+	"encoding/asn1"
 	"errors"
+	"math/big"
 	"strings"
 	"testing"
 
@@ -229,4 +237,143 @@ func TestFindObjectSucceeds(t *testing.T) {
 	handle, err := s.FindObject(nil)
 	test.AssertNotError(t, err, "FindObject failed when everything worked as expected")
 	test.AssertEquals(t, handle, pkcs11.ObjectHandle(1))
+}
+
+func TestGetKey(t *testing.T) {
+	s, ctx := NewSessionWithMock()
+
+	// test newSigner fails when FindObject for private key handle fails
+	ctx.FindObjectsInitFunc = func(pkcs11.SessionHandle, []*pkcs11.Attribute) error {
+		return errors.New("broken")
+	}
+	_, err := s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when FindObject for private key handle failed")
+
+	// test newSigner fails when GetAttributeValue fails
+	ctx.FindObjectsInitFunc = func(pkcs11.SessionHandle, []*pkcs11.Attribute) error {
+		return nil
+	}
+	ctx.FindObjectsFunc = func(pkcs11.SessionHandle, int) ([]pkcs11.ObjectHandle, bool, error) {
+		return []pkcs11.ObjectHandle{1}, false, nil
+	}
+	ctx.FindObjectsFinalFunc = func(pkcs11.SessionHandle) error {
+		return nil
+	}
+	ctx.GetAttributeValueFunc = func(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		return nil, errors.New("broken")
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when GetAttributeValue for private key type failed")
+
+	// test newSigner fails when GetAttributeValue returns no attributes
+	ctx.GetAttributeValueFunc = func(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		return nil, nil
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when GetAttributeValue for private key type returned no attributes")
+
+	// test newSigner fails when FindObject for public key handle fails
+	ctx.GetAttributeValueFunc = func(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		return []*pkcs11.Attribute{pkcs11.NewAttribute(pkcs11.CKA_KEY_TYPE, pkcs11.CKK_EC)}, nil
+	}
+	ctx.FindObjectsInitFunc = func(_ pkcs11.SessionHandle, tmpl []*pkcs11.Attribute) error {
+		if bytes.Equal(tmpl[0].Value, []byte{2, 0, 0, 0, 0, 0, 0, 0}) {
+			return errors.New("broken")
+		}
+		return nil
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when FindObject for public key handle failed")
+
+	// test newSigner fails when FindObject for private key returns unknown CKA_KEY_TYPE
+	ctx.FindObjectsInitFunc = func(_ pkcs11.SessionHandle, tmpl []*pkcs11.Attribute) error {
+		return nil
+	}
+	ctx.GetAttributeValueFunc = func(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		return []*pkcs11.Attribute{pkcs11.NewAttribute(pkcs11.CKA_KEY_TYPE, []byte{2, 0, 0, 0, 0, 0, 0, 0})}, nil
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when GetAttributeValue for private key returned unknown key type")
+
+	// test newSigner fails when GetRSAPublicKey fails
+	ctx.GetAttributeValueFunc = func(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		return []*pkcs11.Attribute{pkcs11.NewAttribute(pkcs11.CKA_KEY_TYPE, []byte{0, 0, 0, 0, 0, 0, 0, 0})}, nil
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when GetRSAPublicKey fails")
+
+	// test newSigner fails when GetECDSAPublicKey fails
+	ctx.GetAttributeValueFunc = func(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		return []*pkcs11.Attribute{pkcs11.NewAttribute(pkcs11.CKA_KEY_TYPE, []byte{3, 0, 0, 0, 0, 0, 0, 0})}, nil
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertError(t, err, "newSigner didn't fail when GetECDSAPublicKey fails")
+
+	// test newSigner works when everything... works
+	ctx.GetAttributeValueFunc = func(_ pkcs11.SessionHandle, _ pkcs11.ObjectHandle, attrs []*pkcs11.Attribute) ([]*pkcs11.Attribute, error) {
+		var returns []*pkcs11.Attribute
+		for _, attr := range attrs {
+			switch attr.Type {
+			case pkcs11.CKA_KEY_TYPE:
+				returns = append(returns, pkcs11.NewAttribute(pkcs11.CKA_KEY_TYPE, []byte{0, 0, 0, 0, 0, 0, 0, 0}))
+			case pkcs11.CKA_PUBLIC_EXPONENT:
+				returns = append(returns, pkcs11.NewAttribute(pkcs11.CKA_PUBLIC_EXPONENT, []byte{1, 2, 3}))
+			case pkcs11.CKA_MODULUS:
+				returns = append(returns, pkcs11.NewAttribute(pkcs11.CKA_MODULUS, []byte{4, 5, 6}))
+			default:
+				return nil, errors.New("GetAttributeValue got unexpected attribute type")
+			}
+		}
+		return returns, nil
+	}
+	_, err = s.NewSigner("label", []byte{255, 255})
+	test.AssertNotError(t, err, "newSigner failed when everything worked properly")
+}
+
+func TestX509Signer(t *testing.T) {
+	s, ctx := NewSessionWithMock()
+
+	// test that x509Signer.Sign properly converts the PKCS#11 format signature to
+	// the RFC 5480 format signature
+	ctx.SignInitFunc = func(pkcs11.SessionHandle, []*pkcs11.Mechanism, pkcs11.ObjectHandle) error {
+		return nil
+	}
+	tk, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	test.AssertNotError(t, err, "Failed to generate test key")
+	ctx.SignFunc = func(_ pkcs11.SessionHandle, digest []byte) ([]byte, error) {
+		r, s, err := ecdsa.Sign(rand.Reader, tk, digest[:])
+		if err != nil {
+			return nil, err
+		}
+		rBytes := r.Bytes()
+		sBytes := s.Bytes()
+		// http://docs.oasis-open.org/pkcs11/pkcs11-curr/v2.40/os/pkcs11-curr-v2.40-os.html
+		// Section 2.3.1: EC Signatures
+		// "If r and s have different octet length, the shorter of both must be padded with
+		// leading zero octets such that both have the same octet length."
+		switch {
+		case len(rBytes) < len(sBytes):
+			padding := make([]byte, len(sBytes)-len(rBytes))
+			rBytes = append(padding, rBytes...)
+		case len(rBytes) > len(sBytes):
+			padding := make([]byte, len(rBytes)-len(sBytes))
+			sBytes = append(padding, sBytes...)
+		}
+		return append(rBytes, sBytes...), nil
+	}
+	digest := sha256.Sum256([]byte("hello"))
+	signer := &X509Signer{session: s, keyType: ECDSAKey, pub: tk.Public()}
+	signature, err := signer.Sign(nil, digest[:], crypto.SHA256)
+	test.AssertNotError(t, err, "x509Signer.Sign failed")
+
+	var rfcFormat struct {
+		R, S *big.Int
+	}
+	rest, err := asn1.Unmarshal(signature, &rfcFormat)
+	test.AssertNotError(t, err, "asn1.Unmarshal failed trying to parse signature")
+	test.Assert(t, len(rest) == 0, "Signature had trailing garbage")
+	verified := ecdsa.Verify(&tk.PublicKey, digest[:], rfcFormat.R, rfcFormat.S)
+	test.Assert(t, verified, "Failed to verify RFC format signature")
+	// For the sake of coverage
+	test.AssertEquals(t, signer.Public(), tk.Public())
 }
